python/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
montydb==2.5.5
numpy==1.26.4
torch==2.2.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
sentence-transformers>=2.7.0
transformers>=4.41.0
openai-whisper>=20231117
omegaconf>=2.3.0
soundfile>=0.12.1
dependency-injector>=4.41.0
